requirements.txt
fastapi
uvicorn
bcrypt>=4.0.1
pydantic-settings
python-dotenv
pytest
pytest-asyncio
pytest-cov
httpx
requests
PyJWT
passlib[bcrypt]